+++ spec/shelly/cli/main_spec.rb
@@ -1101,14 +1101,38 @@ Wait until cloud is in 'turned off' state and try again.")
             invoke(@main, :setup)
           end
         end
+
+        it "should show info about adding default remote and fetching changes" do
+          $stdout.should_receive(:puts).with(green "Setting up foo-staging cloud")
+          $stdout.should_receive(:puts).with("git remote add shelly git_url")
+          $stdout.should_receive(:puts).with("git fetch shelly")
+          $stdout.should_receive(:puts).with(green "Your application is set up.")
+          fake_stdin(["yes"]) do
+            invoke(@main, :setup)
+          end
+        end
       end
 
       context "and user answers no" do
-        it "should display commands to perform manually" do
+        before do
           @app.stub(:git_remote_exist?).with('remote').and_return(false)
+        end
+
+        it "should display commands to perform manually" do
           $stdout.should_receive(:print).with("Specify remote name: ")
           @app.should_receive(:add_git_remote).with('remote')
           @app.should_receive(:git_fetch_remote).with('remote')
+          fake_stdin(["no", "remote"]) do
+            invoke(@main, :setup)
+          end
+        end
+
+        it "should show info about adding custom remote and fetching changes" do
+          $stdout.should_receive(:puts).with(green "Setting up foo-staging cloud")
+          $stdout.should_receive(:print).with("Specify remote name: ")
+          $stdout.should_receive(:puts).with("git remote add remote git_url")
+          $stdout.should_receive(:puts).with("git fetch remote")
+          $stdout.should_receive(:puts).with(green "Your application is set up.")
           fake_stdin(["no", "remote"]) do
             invoke(@main, :setup)
           end