+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [bugfix] Do not show `shelly deploys show last` instruction for last failed internal deployment.
+
 ## 0.4.2 / 2013-08-29
 
 * [improvement] Use thor version without binaries