+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## master
+
+* [bugfix] Do not show `shelly deploys show last` instruction if last deployment was made by admin and was failed.
+* [improvement] There is no longer configuration_failed state for apps.
+
 ## 0.4.2 / 2013-08-29
 
 * [improvement] Use thor version without binaries