+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## master
+## 0.4.3 / 2013-09-10
 
 * [bugfix] Do not show `shelly deploys show last` instruction if last deployment was made by admin and was failed.
 * [improvement] There is no longer configuration_failed state for apps.